requirements.txt
pygame-ce
streamlit